Split and refactor visit utils
