+++ server/routes/visit/update/isPublicBooking.ts
@@ -1,0 +1,5 @@
+import type { EventAudit } from '../../../data/orchestrationApiTypes'
+
+export default (events: EventAudit[]): boolean => {
+  return events.find(event => event.type === 'BOOKED_VISIT')?.userType === 'PUBLIC'
+}
