+++ server/routes/visit/visitUtils.test.ts
@@ -1,23 +1,5 @@
-import { MoJAlert } from '../../@types/bapv'
-import config from '../../config'
-import { notificationTypeAlerts } from '../../constants/notifications'
-import { visitCancellationAlerts } from '../../constants/visitCancellation'
-import { EventAudit, VisitBookingDetails } from '../../data/orchestrationApiTypes'
-import TestData from '../testutils/testData'
-import {
-  getPrisonerLocation,
-  getAvailableVisitActions,
-  AvailableVisitActions,
-  isPublicBooking,
-  getVisitAlerts,
-  getIdsToFlag,
-  getHideAlertsInset,
-} from './visitUtils'
-
-beforeEach(() => {
-  jest.restoreAllMocks()
-  jest.useRealTimers()
-})
+import { VisitBookingDetails } from '../../data/orchestrationApiTypes'
+import { getPrisonerLocation, getIdsToFlag } from './visitUtils'
 
 describe('Visit utils', () => {
   describe('getPrisonerLocation', () => {
@@ -55,398 +37,6 @@ describe('Visit utils', () => {
     })
   })
 
-  describe('getAvailableVisitActions', () => {
-    const startTimestamp = '2025-04-01T09:00:00'
-    let params: Parameters<typeof getAvailableVisitActions>[number]
-
-    beforeEach(() => {
-      params = { visitStatus: 'BOOKED', visitSubStatus: 'AUTO_APPROVED', startTimestamp, notifications: [] }
-    })
-
-    describe('REQUESTED visit', () => {
-      it('should enable only "processRequest" action if visit REQUESTED', () => {
-        params.visitSubStatus = 'REQUESTED'
-
-        expect(getAvailableVisitActions(params)).toStrictEqual<AvailableVisitActions>({
-          update: false,
-          cancel: false,
-          clearNotifications: false,
-          print: false,
-          processRequest: true,
-        })
-      })
-    })
-
-    describe('CANCELLED visit', () => {
-      it('should set all actions to false if visit is CANCELLED', () => {
-        params.visitStatus = 'CANCELLED'
-
-        expect(getAvailableVisitActions(params)).toStrictEqual<AvailableVisitActions>({
-          update: false,
-          cancel: false,
-          clearNotifications: false,
-          print: false,
-          processRequest: false,
-        })
-      })
-    })
-
-    describe('BOOKED visit', () => {
-      describe('update action', () => {
-        it('should set update to false and print to true if visit has started', () => {
-          jest.useFakeTimers({ now: new Date('2025-04-01T09:00:00') })
-          expect(getAvailableVisitActions(params).update).toBe(false)
-          expect(getAvailableVisitActions(params).print).toBe(true)
-        })
-
-        it('should set update and print to true before visit has started', () => {
-          jest.useFakeTimers({ now: new Date('2025-04-01T08:59:59') })
-          expect(getAvailableVisitActions(params).update).toBe(true)
-          expect(getAvailableVisitActions(params).print).toBe(true)
-        })
-
-        it('should set update to false if before start time but prisoner has been released', () => {
-          jest.useFakeTimers({ now: new Date('2025-04-01T08:59:59') })
-          params.notifications = [{ type: 'PRISONER_RELEASED_EVENT' }] as VisitBookingDetails['notifications']
-          expect(getAvailableVisitActions(params).update).toBe(false)
-        })
-
-        it('should set update to false if before start time but prisoner has been transferred', () => {
-          jest.useFakeTimers({ now: new Date('2025-04-01T08:59:59') })
-          params.notifications = [{ type: 'PRISONER_RECEIVED_EVENT' }] as VisitBookingDetails['notifications']
-          expect(getAvailableVisitActions(params).update).toBe(false)
-        })
-      })
-
-      describe('cancel action', () => {
-        it('should set cancel and print to true before visit start time', () => {
-          jest.useFakeTimers({ now: new Date('2025-04-01T08:59:59') })
-          expect(getAvailableVisitActions(params).cancel).toBe(true)
-          expect(getAvailableVisitActions(params).print).toBe(true)
-        })
-
-        it('should set cancel to true and print to false after visit start time, up to CANCELLATION_LIMIT_DAYS (28)', () => {
-          jest.useFakeTimers({ now: new Date('2025-04-29T08:59:59') }) // startTimestamp + 28 days
-          expect(getAvailableVisitActions(params).cancel).toBe(true)
-          expect(getAvailableVisitActions(params).print).toBe(false)
-        })
-
-        it('should set cancel to false and print to false after CANCELLATION_LIMIT_DAYS (28)', () => {
-          jest.useFakeTimers({ now: new Date('2025-04-29T09:00:00') }) // startTimestamp + 28 days + 1 second
-          expect(getAvailableVisitActions(params).cancel).toBe(false)
-          expect(getAvailableVisitActions(params).print).toBe(false)
-        })
-      })
-
-      describe('clearNotifications action', () => {
-        it('should set clearNotifications to false if no visit notifications', () => {
-          expect(getAvailableVisitActions(params).clearNotifications).toBe(false)
-        })
-
-        it('should set clearNotifications to true if a visit notification is set', () => {
-          params.notifications = [{ type: 'NON_ASSOCIATION_EVENT' }] as VisitBookingDetails['notifications']
-          expect(getAvailableVisitActions(params).clearNotifications).toBe(true)
-        })
-
-        it('should set clearNotifications to false if visit notifications includes PRISON_VISITS_BLOCKED_FOR_DATE', () => {
-          params.notifications = [
-            { type: 'NON_ASSOCIATION_EVENT' },
-            { type: 'PRISON_VISITS_BLOCKED_FOR_DATE' },
-          ] as VisitBookingDetails['notifications']
-          expect(getAvailableVisitActions(params).clearNotifications).toBe(false)
-        })
-
-        it('should set clearNotifications to false if visit notifications includes SESSION_VISITS_BLOCKED_FOR_DATE', () => {
-          params.notifications = [
-            { type: 'NON_ASSOCIATION_EVENT' },
-            { type: 'SESSION_VISITS_BLOCKED_FOR_DATE' },
-          ] as VisitBookingDetails['notifications']
-          expect(getAvailableVisitActions(params).clearNotifications).toBe(false)
-        })
-
-        it('should set clearNotifications to false if visit notifications includes VISITOR_UNAPPROVED_EVENT', () => {
-          params.notifications = [
-            { type: 'NON_ASSOCIATION_EVENT' },
-            { type: 'VISITOR_UNAPPROVED_EVENT' },
-          ] as VisitBookingDetails['notifications']
-          expect(getAvailableVisitActions(params).clearNotifications).toBe(false)
-        })
-
-        it('should set print to false if a visit notification is set', () => {
-          params.notifications = [{ type: 'NON_ASSOCIATION_EVENT' }] as VisitBookingDetails['notifications']
-          expect(getAvailableVisitActions(params).print).toBe(false)
-        })
-      })
-    })
-  })
-
-  describe('getVisitAlerts', () => {
-    describe('Visit CANCELLED', () => {
-      it('should return no alert if visitStatus **not** CANCELLED', () => {
-        const visitDetails = TestData.visitBookingDetails({
-          visitStatus: 'BOOKED',
-          visitSubStatus: 'AUTO_APPROVED',
-          outcomeStatus: 'ADMINISTRATIVE_CANCELLATION',
-        })
-        expect(getVisitAlerts(visitDetails)).toStrictEqual([])
-      })
-
-      describe('cancellation reasons', () => {
-        it.each([
-          ['an expected outcomeStatus value', 'BOOKER_CANCELLED', visitCancellationAlerts.BOOKER_CANCELLED],
-          ['fallback to default', '** ANY OTHER STATUS **', visitCancellationAlerts.default],
-          ['handle empty string', '', visitCancellationAlerts.default],
-          ['handle undefined', undefined, visitCancellationAlerts.default],
-        ])(
-          "should handle %s: '%s' => %s",
-          (_: string, outcomeStatus: VisitBookingDetails['outcomeStatus'], expected: string) => {
-            const visitDetails = TestData.visitBookingDetails({
-              visitStatus: 'CANCELLED',
-              visitSubStatus: 'CANCELLED',
-              outcomeStatus,
-            })
-            expect(getVisitAlerts(visitDetails)).toStrictEqual<MoJAlert[]>([
-              {
-                variant: 'information',
-                title: 'Visit cancelled',
-                showTitleAsHeading: true,
-                text: expected,
-              },
-            ])
-          },
-        )
-      })
-    })
-
-    describe('Visit request REJECTED', () => {
-      it('should return an alert if visit request is REJECTED', () => {
-        const visitDetails = TestData.visitBookingDetails({
-          visitStatus: 'CANCELLED',
-          visitSubStatus: 'REJECTED',
-          outcomeStatus: 'ESTABLISHMENT_CANCELLED',
-          events: [
-            {
-              type: 'REQUESTED_VISIT_REJECTED',
-              applicationMethodType: 'WEBSITE',
-              actionedByFullName: 'User One',
-              userType: 'STAFF',
-              createTimestamp: '',
-            },
-          ],
-        })
-
-        expect(getVisitAlerts(visitDetails)).toStrictEqual<MoJAlert[]>([
-          {
-            variant: 'information',
-            title: 'Request rejected',
-            showTitleAsHeading: true,
-            text: 'This visit request was rejected by User One',
-          },
-        ])
-      })
-
-      it('should return an alert if visit request is WITHDRAWN', () => {
-        const visitDetails = TestData.visitBookingDetails({
-          visitStatus: 'CANCELLED',
-          visitSubStatus: 'WITHDRAWN',
-          outcomeStatus: 'REQUESTED_VISIT_WITHDRAWN',
-          events: [
-            {
-              type: 'REQUESTED_VISIT_WITHDRAWN',
-              applicationMethodType: 'WEBSITE',
-              actionedByFullName: undefined,
-              userType: 'PUBLIC',
-              createTimestamp: '',
-            },
-          ],
-        })
-
-        expect(getVisitAlerts(visitDetails)).toStrictEqual<MoJAlert[]>([
-          {
-            variant: 'information',
-            title: 'Request withdrawn',
-            showTitleAsHeading: true,
-            text: 'This visit request was withdrawn by the booker',
-          },
-        ])
-      })
-    })
-
-    describe('Visit request', () => {
-      it('should return no alert if visit is *not* in REQUESTED state', () => {
-        const visitDetails = TestData.visitBookingDetails({ visitSubStatus: 'APPROVED' })
-        expect(getVisitAlerts(visitDetails)).toStrictEqual([])
-      })
-
-      it('should return alert component for a REQUESTED visit', () => {
-        const visitDetails = TestData.visitBookingDetails({ visitSubStatus: 'REQUESTED' })
-        expect(getVisitAlerts(visitDetails)[0].title).toBe('This request needs to be reviewed')
-      })
-    })
-
-    describe('Visit notifications', () => {
-      describe('Single notification(s) => single alert(s)', () => {
-        it.each([
-          [
-            'an expected notification type',
-            [{ type: 'PRISONER_RELEASED_EVENT' }],
-            [notificationTypeAlerts.PRISONER_RELEASED_EVENT],
-          ],
-          [
-            'multiple expected notifications',
-            [{ type: 'PRISONER_RELEASED_EVENT' }, { type: 'PRISON_VISITS_BLOCKED_FOR_DATE' }],
-            [notificationTypeAlerts.PRISONER_RELEASED_EVENT, notificationTypeAlerts.PRISON_VISITS_BLOCKED_FOR_DATE],
-          ],
-          [
-            'an unexpected notification type',
-            [{ type: 'PRISONER_RELEASED_EVENT' }, { type: '** SOME OTHER TYPE **' }],
-            [notificationTypeAlerts.PRISONER_RELEASED_EVENT],
-          ],
-          ['no notifications', [], []],
-        ])(
-          'should handle %s',
-          (_: string, notifications: VisitBookingDetails['notifications'], expected: MoJAlert[]) => {
-            const visitDetails = TestData.visitBookingDetails({ notifications })
-            expect(getVisitAlerts(visitDetails)).toStrictEqual(expected)
-          },
-        )
-      })
-
-      describe('Mixed notifications', () => {
-        it.each([
-          [
-            'a single visitor restriction notification',
-            [
-              {
-                type: 'VISITOR_RESTRICTION',
-                additionalData: [{ attributeName: 'VISITOR_RESTRICTION_ID', attributeValue: '1' }],
-              },
-            ],
-            [
-              {
-                variant: 'warning',
-                title: 'This visit needs review',
-                showTitleAsHeading: true,
-                html:
-                  '<ul class="govuk-list">' +
-                  '<li><a href="#visitor-restriction-1">A restriction has been added or updated</a></li>' +
-                  '</ul>',
-                classes: 'notifications-summary-alert',
-              },
-            ] as MoJAlert[],
-          ],
-
-          [
-            'two visitor restriction notifications',
-            [
-              {
-                type: 'VISITOR_RESTRICTION',
-                additionalData: [{ attributeName: 'VISITOR_RESTRICTION_ID', attributeValue: '1' }],
-              },
-              // a duplicate VISITOR_RESTRICTION_ID - should be ignored
-              {
-                type: 'VISITOR_RESTRICTION',
-                additionalData: [{ attributeName: 'VISITOR_RESTRICTION_ID', attributeValue: '1' }],
-              },
-              {
-                type: 'VISITOR_RESTRICTION',
-                additionalData: [{ attributeName: 'VISITOR_RESTRICTION_ID', attributeValue: '2' }],
-              },
-            ],
-            [
-              {
-                variant: 'warning',
-                title: 'This visit needs review',
-                showTitleAsHeading: true,
-                html:
-                  '<ul class="govuk-list">' +
-                  '<li><a href="#visitor-restriction-1">A restriction has been added or updated</a></li>' +
-                  '<li><a href="#visitor-restriction-2">A restriction has been added or updated</a></li>' +
-                  '</ul>',
-                classes: 'notifications-summary-alert',
-              },
-            ],
-          ],
-          [
-            'visitor restriction and visitor unapproved notifications',
-            [
-              {
-                type: 'VISITOR_RESTRICTION',
-                additionalData: [{ attributeName: 'VISITOR_RESTRICTION_ID', attributeValue: '2' }],
-              },
-              {
-                type: 'VISITOR_UNAPPROVED_EVENT',
-                additionalData: [{ attributeName: 'VISITOR_ID', attributeValue: '3' }],
-              },
-              {
-                type: 'PRISONER_ALERT_UPDATED_EVENT',
-                additionalData: [{ attributeName: 'ALERT_UUID', attributeValue: '4' }],
-              },
-              {
-                type: 'PRISONER_ALERT_CREATED_EVENT',
-                additionalData: [{ attributeName: 'ALERT_UUID', attributeValue: '5' }],
-              },
-            ],
-            [
-              {
-                variant: 'warning',
-                title: 'This visit needs review',
-                showTitleAsHeading: true,
-                html:
-                  '<ul class="govuk-list">' +
-                  '<li><a href="#visitor-restriction-2">A restriction has been added or updated</a></li>' +
-                  '<li><a href="#visitor-3">Visitor has been unapproved</a></li>' +
-                  '<li><a href="#prisoner-alert-4">An alert has been updated</a></li>' +
-                  '<li><a href="#prisoner-alert-5">An alert has been added</a></li>' +
-                  '</ul>',
-                classes: 'notifications-summary-alert',
-              },
-            ],
-          ],
-        ])(
-          'should handle %s',
-          (_: string, notifications: VisitBookingDetails['notifications'], expected: MoJAlert[]) => {
-            const visitDetails = TestData.visitBookingDetails({ notifications })
-            expect(getVisitAlerts(visitDetails)).toStrictEqual(expected)
-          },
-        )
-      })
-
-      describe('Mixed single/grouped notifications', () => {
-        it('should handle a mix of notifications: one single and two to group as one', () => {
-          const notifications = [
-            { type: 'PRISONER_RELEASED_EVENT' },
-            {
-              type: 'VISITOR_RESTRICTION',
-              additionalData: [{ attributeName: 'VISITOR_RESTRICTION_ID', attributeValue: '1' }],
-            },
-            {
-              type: 'VISITOR_RESTRICTION',
-              additionalData: [{ attributeName: 'VISITOR_RESTRICTION_ID', attributeValue: '2' }],
-            },
-          ] as VisitBookingDetails['notifications']
-
-          const visitDetails = TestData.visitBookingDetails({ notifications })
-          const result = getVisitAlerts(visitDetails)
-
-          expect(result).toStrictEqual<MoJAlert[]>([
-            notificationTypeAlerts.PRISONER_RELEASED_EVENT,
-            {
-              variant: 'warning',
-              title: 'This visit needs review',
-              showTitleAsHeading: true,
-              html:
-                '<ul class="govuk-list">' +
-                '<li><a href="#visitor-restriction-1">A restriction has been added or updated</a></li>' +
-                '<li><a href="#visitor-restriction-2">A restriction has been added or updated</a></li>' +
-                '</ul>',
-              classes: 'notifications-summary-alert',
-            } as MoJAlert,
-          ])
-        })
-      })
-    })
-  })
   describe('getIdsToFlag', () => {
     // flaggedVisitorRestrictionIds
     const visitorRestrictionId = 'VISITOR_RESTRICTION_ID'
@@ -558,96 +148,6 @@ describe('Visit utils', () => {
       expect(
         getIdsToFlag({ notificationType: alertCreatedEvent, returnedIdType: alertUuid, notifications }),
       ).toStrictEqual(['100', '200'])
-    })
-  })
-
-  describe('isPublicBooking', () => {
-    const events: EventAudit[] = [
-      {
-        applicationMethodType: 'NOT_KNOWN',
-        createTimestamp: '2022-01-01T09:00:00',
-        type: 'UPDATED_VISIT',
-        userType: 'STAFF',
-      },
-    ]
-
-    it('should return true if visit booked event is from public user type', () => {
-      events[0].type = 'BOOKED_VISIT'
-      events[0].userType = 'PUBLIC'
-
-      const publicBooker = isPublicBooking(events)
-      expect(publicBooker).toStrictEqual(true)
-    })
-
-    it('should return false if visit booked event is from staff user type', () => {
-      events[0].type = 'BOOKED_VISIT'
-      events[0].userType = 'STAFF'
-
-      const publicBooker = isPublicBooking(events)
-      expect(publicBooker).toStrictEqual(false)
-    })
-
-    it('should return false if no visit booked event is present', () => {
-      events[0].type = 'UPDATED_VISIT'
-      events[0].userType = 'PUBLIC'
-
-      const publicBooker = isPublicBooking(events)
-      expect(publicBooker).toStrictEqual(false)
-    })
-  })
-
-  describe('getHideAlertsInset', () => {
-    const prisonerNumber = 'A1234BC'
-    let expected: ReturnType<typeof getHideAlertsInset>
-
-    beforeEach(() => {
-      expected = {
-        prisoner: {
-          attributes: { 'data-test': 'prisoner-inset' },
-          classes: 'govuk-!-margin-bottom-1',
-        },
-        visitor: {
-          attributes: { 'data-test': 'visitor-inset' },
-        },
-      } as ReturnType<typeof getHideAlertsInset>
-    })
-
-    it('should return past visit message if skipAlertsAndRestrictionReason is VISIT_IN_PAST', () => {
-      expected.prisoner.html = `Alerts and restrictions are not shown for past visits.<br>You can view alerts and restrictions in the <a href="${config.dpsContacts}/prisoner/${prisonerNumber}/alerts-restrictions">contacts service</a>.`
-      expected.visitor.html = `Visitor restrictions are not shown for past visits.<br>You can view restrictions in the <a href="${config.dpsContacts}/prisoner/${prisonerNumber}/contacts/list">contacts service</a>.`
-
-      const results = getHideAlertsInset({ skipAlertsAndRestrictionReason: 'VISIT_IN_PAST', prisonerNumber })
-      expect(results).toStrictEqual(expected)
-    })
-
-    it('should return released prisoner message if skipAlertsAndRestrictionReason is PRISONER_RELEASED', () => {
-      expected.prisoner.text = 'Alerts and restrictions are not shown for released prisoners.'
-      expected.visitor.text = 'Visitor restrictions are not shown for released prisoners.'
-
-      const results = getHideAlertsInset({
-        skipAlertsAndRestrictionReason: 'PRISONER_RELEASED',
-        prisonerNumber,
-      })
-      expect(results).toStrictEqual(expected)
-    })
-
-    it('should return transferred prisoner message if skipAlertsAndRestrictionReason is PRISONER_TRANSFERRED', () => {
-      expected.prisoner.text = 'Alerts and restrictions are not shown for transferred prisoners.'
-      expected.visitor.text = 'Visitor restrictions are not shown for transferred prisoners.'
-
-      const results = getHideAlertsInset({
-        prisonerNumber,
-        skipAlertsAndRestrictionReason: 'PRISONER_TRANSFERRED',
-      })
-      expect(results).toStrictEqual(expected)
-    })
-
-    it('should return null if skipAlertsAndRestrictionReason is null', () => {
-      const results = getHideAlertsInset({
-        prisonerNumber,
-        skipAlertsAndRestrictionReason: null,
-      })
-      expect(results).toStrictEqual(null)
     })
   })
 })
