+++ server/routes/visit/visitUtils.ts
@@ -1,17 +1,8 @@
-import { isFuture, isToday } from 'date-fns'
-import { GOVUKInsetText, MoJAlert, TextOrHtml } from '../../@types/bapv'
-import config from '../../config'
-import { notificationTypeAlerts } from '../../constants/notifications'
-import { visitCancellationAlerts } from '../../constants/visitCancellation'
 import {
-  EventAudit,
   NotificationType,
   VisitBookingDetails,
   VisitNotificationEventAttributeNames,
 } from '../../data/orchestrationApiTypes'
-
-const A_DAY_IN_MS = 24 * 60 * 60 * 1000
-const CANCELLATION_LIMIT_MS = config.visit.cancellationLimitDays * A_DAY_IN_MS
 
 export const getPrisonerLocation = (prisoner: VisitBookingDetails['prisoner']) => {
   if (prisoner.prisonId === 'OUT') {
@@ -23,240 +14,6 @@ export const getPrisonerLocation = (prisoner: VisitBookingDetails['prisoner']) =
   }
 
   return `${prisoner.cellLocation}, ${prisoner.prisonName}`
-}
-
-export type AvailableVisitActions = ReturnType<typeof getAvailableVisitActions>
-
-export const getAvailableVisitActions = ({
-  visitStatus,
-  visitSubStatus,
-  startTimestamp,
-  notifications,
-}: {
-  visitStatus: VisitBookingDetails['visitStatus']
-  visitSubStatus: VisitBookingDetails['visitSubStatus']
-  startTimestamp: VisitBookingDetails['startTimestamp']
-  notifications: VisitBookingDetails['notifications']
-}): { update: boolean; cancel: boolean; clearNotifications: boolean; print: boolean; processRequest: boolean } => {
-  const availableVisitActions = {
-    update: false,
-    cancel: false,
-    clearNotifications: false,
-    print: false,
-    processRequest: false,
-  }
-
-  if (visitSubStatus === 'REQUESTED') {
-    availableVisitActions.processRequest = true
-    return availableVisitActions
-  }
-
-  if (visitStatus !== 'BOOKED') {
-    return availableVisitActions
-  }
-
-  const visitStartTime = new Date(startTimestamp)
-
-  // update
-  const hasUpdateBlockingNotifications = notifications.some(
-    notification => notification.type === 'PRISONER_RECEIVED_EVENT' || notification.type === 'PRISONER_RELEASED_EVENT',
-  )
-
-  if (!hasUpdateBlockingNotifications && isFuture(visitStartTime)) {
-    availableVisitActions.update = true
-  }
-
-  // print
-  if (!notifications.length && (isToday(visitStartTime) || isFuture(visitStartTime))) {
-    availableVisitActions.print = true
-  }
-
-  // cancel
-  const latestCancellationTime = new Date(visitStartTime.getTime() + CANCELLATION_LIMIT_MS)
-
-  if (isFuture(latestCancellationTime)) {
-    availableVisitActions.cancel = true
-  }
-
-  // clear notifications ('do not change')
-  const hasBlockedDateOrSessionNotification = notifications.some(
-    notification =>
-      notification.type === 'PRISON_VISITS_BLOCKED_FOR_DATE' || notification.type === 'SESSION_VISITS_BLOCKED_FOR_DATE',
-  )
-
-  const hasUnapprovedVisitorNotification = notifications.some(
-    notification => notification.type === 'VISITOR_UNAPPROVED_EVENT',
-  )
-
-  if (!hasBlockedDateOrSessionNotification && !hasUnapprovedVisitorNotification && notifications.length > 0) {
-    availableVisitActions.clearNotifications = true
-  }
-
-  return availableVisitActions
-}
-
-const getVisitCancelledAlert = ({
-  visitStatus,
-  visitSubStatus,
-  outcomeStatus,
-}: {
-  visitStatus: VisitBookingDetails['visitStatus']
-  visitSubStatus: VisitBookingDetails['visitSubStatus']
-  outcomeStatus: VisitBookingDetails['outcomeStatus']
-}): MoJAlert | undefined => {
-  if (visitStatus !== 'CANCELLED' || visitSubStatus !== 'CANCELLED') {
-    return undefined
-  }
-
-  return {
-    variant: 'information',
-    title: 'Visit cancelled',
-    showTitleAsHeading: true,
-    text: visitCancellationAlerts[outcomeStatus] ?? visitCancellationAlerts.default,
-  }
-}
-
-const getVisitRequestAlert = ({
-  visitStatus,
-  visitSubStatus,
-  events,
-}: {
-  visitStatus: VisitBookingDetails['visitStatus']
-  visitSubStatus: VisitBookingDetails['visitSubStatus']
-  events: VisitBookingDetails['events']
-}): MoJAlert | undefined => {
-  if (visitSubStatus === 'REQUESTED') {
-    return {
-      variant: 'warning',
-      title: 'This request needs to be reviewed',
-      showTitleAsHeading: true,
-      text: 'Check alerts and restrictions.',
-    }
-  }
-
-  if (visitStatus === 'CANCELLED' && visitSubStatus === 'REJECTED') {
-    const username = events.find(event => event.type === 'REQUESTED_VISIT_REJECTED')?.actionedByFullName
-
-    return {
-      variant: 'information',
-      title: 'Request rejected',
-      showTitleAsHeading: true,
-      text: `This visit request was rejected by ${username}`,
-    }
-  }
-
-  if (visitStatus === 'CANCELLED' && visitSubStatus === 'WITHDRAWN') {
-    return {
-      variant: 'information',
-      title: 'Request withdrawn',
-      showTitleAsHeading: true,
-      text: `This visit request was withdrawn by the booker`,
-    }
-  }
-
-  return undefined
-}
-
-const getVisitNotificationsAlerts = (notifications: VisitBookingDetails['notifications']): MoJAlert[] => {
-  // simpleNotifications are simple descriptive notifications
-  // linkedNotifications are notifications with anchors to restrictions/visitors on the page
-  const simpleNotifications: VisitBookingDetails['notifications'] = []
-  const linkedNotifications: VisitBookingDetails['notifications'] = []
-  notifications.forEach(notification => {
-    if (
-      notification.type === 'VISITOR_RESTRICTION' ||
-      notification.type === 'VISITOR_UNAPPROVED_EVENT' ||
-      notification.type === 'PRISONER_ALERT_UPDATED_EVENT' ||
-      notification.type === 'PRISONER_ALERT_CREATED_EVENT'
-    ) {
-      linkedNotifications.push(notification)
-    } else {
-      simpleNotifications.push(notification)
-    }
-  })
-
-  if (!simpleNotifications.length && !linkedNotifications.length) {
-    return []
-  }
-
-  const alerts = []
-
-  simpleNotifications.forEach(notification => {
-    if (notificationTypeAlerts[notification.type]) {
-      alerts.push(notificationTypeAlerts[notification.type])
-    }
-  })
-
-  if (linkedNotifications.length) {
-    const visitorRestrictionIds = getIdsToFlag({
-      notificationType: 'VISITOR_RESTRICTION',
-      returnedIdType: 'VISITOR_RESTRICTION_ID',
-      notifications: linkedNotifications,
-    })
-    const unapprovedVisitorIds = getIdsToFlag({
-      notificationType: 'VISITOR_UNAPPROVED_EVENT',
-      returnedIdType: 'VISITOR_ID',
-      notifications: linkedNotifications,
-    })
-    const alertUpdatedIds = getIdsToFlag({
-      notificationType: 'PRISONER_ALERT_UPDATED_EVENT',
-      returnedIdType: 'ALERT_UUID',
-      notifications: linkedNotifications,
-    })
-    const alertCreatedIds = getIdsToFlag({
-      notificationType: 'PRISONER_ALERT_CREATED_EVENT',
-      returnedIdType: 'ALERT_UUID',
-      notifications: linkedNotifications,
-    })
-
-    let notificationItems = visitorRestrictionIds
-      .map(id => `<li><a href="#visitor-restriction-${id}">A restriction has been added or updated</a></li>`)
-      .join('')
-
-    notificationItems += unapprovedVisitorIds
-      .map(id => `<li><a href="#visitor-${id}">Visitor has been unapproved</a></li>`)
-      .join('')
-
-    notificationItems += alertUpdatedIds
-      .map(id => `<li><a href="#prisoner-alert-${id}">An alert has been updated</a></li>`)
-      .join('')
-
-    notificationItems += alertCreatedIds
-      .map(id => `<li><a href="#prisoner-alert-${id}">An alert has been added</a></li>`)
-      .join('')
-
-    alerts.push({
-      variant: 'warning',
-      title: 'This visit needs review',
-      showTitleAsHeading: true,
-      html: `<ul class="govuk-list">${notificationItems}</ul>`,
-      classes: 'notifications-summary-alert',
-    } as MoJAlert)
-  }
-
-  return alerts
-}
-
-export const getVisitAlerts = (visitDetails: VisitBookingDetails): MoJAlert[] => {
-  const { visitStatus, visitSubStatus, outcomeStatus, events, notifications } = visitDetails
-
-  const visitCancelledAlert = getVisitCancelledAlert({
-    visitStatus,
-    visitSubStatus,
-    outcomeStatus,
-  })
-  const visitRequestAlert = getVisitRequestAlert({
-    visitStatus,
-    visitSubStatus,
-    events,
-  })
-  const visitNotificationsAlerts = getVisitNotificationsAlerts(notifications)
-
-  return [
-    ...(visitCancelledAlert ? [visitCancelledAlert] : []),
-    ...(visitRequestAlert ? [visitRequestAlert] : []),
-    ...visitNotificationsAlerts,
-  ]
 }
 
 export const getIdsToFlag = ({
@@ -276,68 +33,4 @@ export const getIdsToFlag = ({
       flaggedIds.add(matchedNotifications?.attributeValue)
     })
   return Array.from(flaggedIds)
-}
-
-export const isPublicBooking = (events: EventAudit[]): boolean => {
-  const visitBookedEvent = events.find(event => event.type === 'BOOKED_VISIT')
-  return visitBookedEvent?.userType === 'PUBLIC'
-}
-
-export const getHideAlertsInset = ({
-  skipAlertsAndRestrictionReason,
-  prisonerNumber,
-}: {
-  skipAlertsAndRestrictionReason: VisitBookingDetails['skipAlertsAndRestrictionReason']
-  prisonerNumber: string
-}): { prisoner: GOVUKInsetText; visitor: GOVUKInsetText } | null => {
-  if (!skipAlertsAndRestrictionReason) {
-    return null
-  }
-
-  let prisonerTextOrHtml: TextOrHtml
-  let visitorTextOrHtml: TextOrHtml
-
-  switch (skipAlertsAndRestrictionReason) {
-    case 'VISIT_IN_PAST':
-      prisonerTextOrHtml = {
-        html: `Alerts and restrictions are not shown for past visits.<br>You can view alerts and restrictions in the <a href="${config.dpsContacts}/prisoner/${prisonerNumber}/alerts-restrictions">contacts service</a>.`,
-      }
-      visitorTextOrHtml = {
-        html: `Visitor restrictions are not shown for past visits.<br>You can view restrictions in the <a href="${config.dpsContacts}/prisoner/${prisonerNumber}/contacts/list">contacts service</a>.`,
-      }
-      break
-
-    case 'PRISONER_RELEASED':
-      prisonerTextOrHtml = {
-        text: 'Alerts and restrictions are not shown for released prisoners.',
-      }
-      visitorTextOrHtml = {
-        text: 'Visitor restrictions are not shown for released prisoners.',
-      }
-      break
-
-    case 'PRISONER_TRANSFERRED':
-      prisonerTextOrHtml = {
-        text: 'Alerts and restrictions are not shown for transferred prisoners.',
-      }
-      visitorTextOrHtml = {
-        text: 'Visitor restrictions are not shown for transferred prisoners.',
-      }
-      break
-
-    default:
-      return null
-  }
-
-  return {
-    prisoner: {
-      ...prisonerTextOrHtml,
-      attributes: { 'data-test': 'prisoner-inset' },
-      classes: 'govuk-!-margin-bottom-1',
-    },
-    visitor: {
-      ...visitorTextOrHtml,
-      attributes: { 'data-test': 'visitor-inset' },
-    },
-  }
 }
